requirements.txt
certifi==2020.6.20
pyserial==3.5
pyusb==1.2.1
typing_extensions==4.1.1
